requirements.txt
PySide6>=6.6
PyMuPDF>=1.23
pillow>=10.0
pytesseract>=0.3.10
python-docx>=1.0.0
genanki>=0.13.1
google-generativeai>=0.6.0
